+++ src/extension.ts
@@ -17,42 +17,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 	let openFileInNewWindow = vscode.commands.registerCommand(op.OpenFileInNewWindowCommandId, op.openFileInNewWindow);
 	context.subscriptions.push(openFileInNewWindow);
-
-	let owtcOverrideMethod = vscode.commands.registerCommand("workbench.action.files.openFile", async (arg) => {
-		owtcOpenFileOverride(owtcOverrideMethod);
-	})
-
-	// Method implementation
-	async function owtcOpenFileOverride(methodDisposer: any) {
-		methodDisposer.dispose();	// Abort and Dispose override method
-		
-		if (vscode.workspace.getConfiguration().get("owtc.overrideOpenFileMethod") === true) {
-			// Call override method
-			const options: vscode.OpenDialogOptions = {
-				canSelectMany: false
-			};
-
-			vscode.window.showOpenDialog(options).then(fileUri => {
-				if (fileUri && fileUri[0]) {
-					op.openFileInNewWindow(fileUri[0]);
-				}
-			});
-		}
-		else {
-			// Call original method
-			await vscode.commands.executeCommand("workbench.action.files.openFile");
-		}
-
-		//	re-register override method
-		let owtcOverrideMethod = vscode.commands.registerCommand("workbench.action.files.openFile", async (arg) => {
-			owtcOpenFileOverride(owtcOverrideMethod);
-		})
-		context.subscriptions.push(owtcOverrideMethod);
-	}
-
-	context.subscriptions.push(owtcOverrideMethod);
 }
-
 
 
 // this method is called when your extension is deactivated