+++ src/path.ts
@@ -13,14 +13,47 @@ export function getCodeCommandPath() {
 		return cache[command];
 	}
 
-	if (process.env["PATH"]) {
-		var pathparts = process.env["PATH"].split(path.delimiter);
-		for (var i = 0; i < pathparts.length; i++) {
-			let binpath = path.join(pathparts[i], command);
-			if (fs.existsSync(binpath)) {
-				cache[command] = binpath;
-				return binpath;
-			}
+	let code_path = "";
+	let candidate_paths: string[] = [];
+	// candidate 1
+	code_path = path.join(path.dirname(process.execPath), "bin", command);
+	if (fs.existsSync(code_path)) {
+		candidate_paths.push(code_path);
+	}
+	// candidate 2
+	code_path = path.join(path.dirname(process.execPath), "../Resources/app/bin", command);
+	if (fs.existsSync(code_path)) {
+		candidate_paths.push(code_path);
+	}
+	// candidate 3
+	code_path = path.join(process.cwd(), command);
+	if (fs.existsSync(code_path)) {
+		candidate_paths.push(code_path);
+	}
+	// candidate 4
+	code_path = path.join(process.cwd(), "bin", command);
+	if (fs.existsSync(code_path)) {
+		candidate_paths.push(code_path);
+	}
+
+	//console.log(candidate_paths);
+
+	if (candidate_paths.length === 0) {
+		return null;
+	}
+
+	if (candidate_paths.length === 1) {
+		cache[command] = candidate_paths[0];
+		return candidate_paths[0];
+	}
+
+	// when hit multiple candidates
+	for (const p of candidate_paths) {
+		const contents = fs.readFileSync(p, 'utf-8');
+		if (contents.includes("VSCODE") && contents.includes("ELECTRON")) {
+			// easy check file contents
+			cache[command] = p;
+			return p;
 		}
 	}
 
@@ -31,9 +64,9 @@ export function isCodeCommandAvailable(): boolean {
 	return getCodeCommandPath() !== null;
 }
 
-function correctCommandName(binname: string) {
+function correctCommandName(bin_name: string) {
 	if (process.platform === "win32") {
-		return binname + ".cmd";
+		return bin_name + ".cmd";
 	}
-	return binname;
+	return bin_name;
 }